weather-mcp/pyproject.toml
[tool.poetry]
name = "weekly-weather-mcp"
version = "0.1.0"
description = "A weather forecast MCP server providing 8-day global weather forecasts"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "*.py"}]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.12.4"
mcp-server = ">=0.1.3"
fastmcp = ">=0.4.1"
requests = ">=2.32.0"
pydantic = "^1.10.13"
python-dotenv = "^1.0.0"
langchain-openai = "^0.1.0"
langchain = "^0.1.0"
openai = "^1.10.0"


[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
black = ">=23.0.0"
mypy = ">=1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
